.qoder/skills/usta-data-parser/scripts/requirements.txt
playwright>=1.44.0
mysql-connector-python>=8.4.0
apscheduler>=3.10.4
python-dotenv>=1.0.1
